Migrate action code to actionbar
